feat(wallet): initialize service with security and feign clients
- Created basic Spring Boot application structure.
- Configured Spring Security with JwtAuthenticationFilter.
- Added IdentityClient for token validation.
- Updated pom.xml dependencies and enabled Feign Clients.

+++ wallet-service/src/main/java/com/hk/hk/walletservice/controller/WalletController.java
@@ -17,7 +17,7 @@ public class WalletController {
     private final WalletService walletService;
 
     @GetMapping("/user/{userId}")
-    public ResponseEntity<WalletResponse> getWalletByUserId(@PathVariable Integer userId) {
+    public ResponseEntity<WalletResponse> getWalletByUserId(@PathVariable Long userId) {
         Wallet wallet = walletService.getByUserId(userId);
 
         WalletResponse response = WalletResponse.builder()
