+++ wallet-service/src/main/java/com/hk/hk/walletservice/repository/WalletRepository.java
@@ -4,8 +4,8 @@ import com.hk-fintech.hk.walletservice.entity.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface WalletRepository extends JpaRepository<Wallet, Integer> {
+public interface WalletRepository extends JpaRepository<Wallet, Long> {
 
-    Optional<Wallet> findByUserId(Integer userId);
+    Optional<Wallet> findByUserId(Long userId);
 
 }
